# api-docs.yaml
openapi: 3.0.3
info:
  title: Marketing Project API
  description: |
    Comprehensive API for marketing content processing and analysis.

    This API provides endpoints for:
    - Content analysis and quality assessment
    - Complete marketing pipeline processing
    - Content source management
    - System health monitoring

    ## Content Types
    The API supports various content types:
    - **Blog Posts**: Marketing blog content
    - **Transcripts**: Audio/video transcripts
    - **Release Notes**: Software release documentation

  version: 1.0.0
  contact:
    name: Marketing Project Team
    email: support@marketing-project.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.marketing-project.com
    description: Production server
paths:
...
  /api/v1/system/info:
    get:
      tags:
        - System
      summary: Get System Information
      description: |
        Get system information and configuration.
        Returns system information including version, configuration status, and environment details.
      operationId: getSystemInfo
      responses:
        '200':
          description: System information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SystemInfo'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    APIResponse:
      type: object
      required:
        - success
        - timestamp
      properties:
        success:
          type: boolean
          description: Whether the operation was successful
        message:
          type: string
          description: Response message
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        request_id:
          type: string
          description: Request ID for tracking
    ErrorResponse:
      allOf:
        - $ref: '#/components/schemas/APIResponse'
        - type: object
          required:
            - success
            - error_code
          properties:
            success:
              type: boolean
              default: false
              description: Always false for errors
            error_code:
              type: string
              description: Error code
            error_details:
              type: object
              description: Additional error details

    # Content Models
...
    SystemInfo:
      type: object
      required:
        - service
        - version
        - python_version
        - platform
        - environment
        - configuration
      properties:
        service:
          type: string
          description: Service name
        version:
          type: string
          description: Service version
        python_version:
          type: string
          description: Python version
        platform:
          type: string
          description: Platform information
        environment:
          type: object
          properties:
            debug:
              type: boolean
              description: Debug mode enabled
            log_level:
              type: string
              description: Log level
            template_version:
              type: string
              description: Template version
        configuration:
          type: object
          properties:
            pipeline_loaded:
              type: boolean
              description: Pipeline configuration loaded
            prompts_dir_exists:
              type: boolean
              description: Prompts directory exists
            prompts_dir:
              type: string
              description: Prompts directory path
...